+++ src/Server/Dependencies/Dependencies.ts
@@ -1,0 +1,7 @@
+import {Logger} from "pino"
+import KeyValueStore from "../Persistence/KeyValueStore"
+
+export default interface Dependencies {
+    logger: Logger
+    kv: KeyValueStore
+}