+++ src/Server/Logging/ConsoleLogger.ts
@@ -1,7 +1,0 @@
-import Logger from "./Logger"
-
-export default class ConsoleLogger implements Logger {
-    log(message: string) {
-        console.log(message)
-    }
-}