+++ src/Server/Logging/MockLogger.ts
@@ -1,9 +1,0 @@
-import Logger from "./Logger"
-
-export default class MockLogger implements Logger {
-    public calls: number = 0
-
-    log(message: string) {
-        this.calls += 1
-    }
-}