+++ src/Server/Startup/preflight.ts
@@ -3,6 +3,11 @@ import fs from "fs"
 import {execSync} from "child_process"
 import runShutdownTasks from "../Shutdown/runShutdownTasks"
 
+// Convert unhandled rejections to thrown errors
+process.on('unhandledRejection', function(err) {
+    throw err
+})
+
 // Register shutdown tasks
 process.on('SIGTERM', async () => {
     await runShutdownTasks()
@@ -13,4 +18,5 @@ if (!fs.existsSync(jetpaxHome)) {
 
     // Make sure the directory exists first too, to be safe
     execSync(`mkdir -p ${jetpaxHome}`)
+
 }